added payments using stripe, added pagination, file uploading, changes

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { errorMonitor } = require('nodemailer/lib/xoauth2');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 const {validationResult} = require('express-validator');
 
@@ -122,6 +123,7 @@ exports.postEditProduct = (req, res, next) => {
     product.price = updatedPrice;
     product.description = updatedDesc;
     if(image) {
+      fileHelper.deleteFile(product.imageUrl);
       product.imageUrl = image.path;
     }
     return product.save()
@@ -156,17 +158,25 @@ exports.getProducts = (req, res, next) => {
   });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId)
+  .then(product => {
+    if(!product) {
+      return next(new Error ('Product not found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
   .then(() => {
     console.log('DESTROYED PRODUCT');
-    res.redirect('/admin/products');
+    res.status(200).json({
+      message: 'Success!'
+    });
   })
   .catch(err => {
-    const error = new Error(err)
-    error.httpStatusCode = 500;
-    return next(error);
+    res.status(500).json({
+      message: 'Deleting product failed!'
+    });
   });
-
 };
